Split up crypto.js to improve bundling
