+++ src/transactions/Association.ts
@@ -1,8 +1,8 @@
 import Transaction from "./Transaction";
-import {concatUint8Arrays} from "../utils/concat";
+import {concatByteArray, strToBytes} from "../utils/byte-array";
 import base58 from "../libs/base58";
 import * as convert from "../utils/convert";
-import * as crypto from "../utils/crypto";
+import {keyTypeId} from "../utils/crypto";
 import {IHash, ISigner, ITxJSON} from "../../interfaces";
 import Binary from "../Binary";
 import {default as DataEntry, dictToData} from "./DataEntry";
@@ -41,23 +41,23 @@ export default class Association extends Transaction {
 
 	private dataToBinary(): Uint8Array {
 		return this.data.reduce(
-			(binary: Uint8Array, entry: DataEntry) => concatUint8Arrays(binary, entry.toBinary()),
+			(binary: Uint8Array, entry: DataEntry) => concatByteArray(binary, entry.toBinary()),
 			new Uint8Array
 		);
 	}
 
 	private toBinaryV1(): Uint8Array {
 		const subjectBinary = this.subject
-			? concatUint8Arrays(
+			? concatByteArray(
 				Uint8Array.from([1]),
 				Uint8Array.from(convert.shortToByteArray(this.subject.length)),
 				this.subject,
 			)
 			: Uint8Array.from([0]);
 
-		return concatUint8Arrays(
+		return concatByteArray(
 			Uint8Array.from([this.type, this.version]),
-			Uint8Array.from(crypto.strToBytes(this.chainId)),
+			Uint8Array.from(strToBytes(this.chainId)),
 			base58.decode(this.senderPublicKey),
 			base58.decode(this.recipient),
 			Uint8Array.from(convert.integerToByteArray(this.associationType)),
@@ -68,11 +68,11 @@ export default class Association extends Transaction {
 	}
 
 	private toBinaryV3(): Uint8Array {
-		return concatUint8Arrays(
+		return concatByteArray(
 			Uint8Array.from([this.type, this.version]),
-			Uint8Array.from(crypto.strToBytes(this.chainId)),
+			Uint8Array.from(strToBytes(this.chainId)),
 			Uint8Array.from(convert.longToByteArray(this.timestamp)),
-			Uint8Array.from([crypto.keyTypeId(this.senderKeyType)]),
+			Uint8Array.from([keyTypeId(this.senderKeyType)]),
 			base58.decode(this.senderPublicKey),
 			Uint8Array.from(convert.longToByteArray(this.fee)),
 			base58.decode(this.recipient),
@@ -84,11 +84,11 @@ export default class Association extends Transaction {
 	}
 
 	private toBinaryV4(): Uint8Array {
-		return concatUint8Arrays(
+		return concatByteArray(
 			Uint8Array.from([this.type, this.version]),
-			Uint8Array.from(crypto.strToBytes(this.chainId)),
+			Uint8Array.from(strToBytes(this.chainId)),
 			Uint8Array.from(convert.longToByteArray(this.timestamp)),
-			Uint8Array.from([crypto.keyTypeId(this.senderKeyType)]),
+			Uint8Array.from([keyTypeId(this.senderKeyType)]),
 			base58.decode(this.senderPublicKey),
 			Uint8Array.from(convert.longToByteArray(this.fee)),
 			Uint8Array.from(convert.longToByteArray(this.associationType)),
