+++ src/transactions/Burn.ts
@@ -1,10 +1,9 @@
 import Transaction from "./Transaction";
-import { concatUint8Arrays } from "../utils/concat";
+import {concatByteArray, strToBytes} from "../utils/byte-array";
 import base58 from "../libs/base58";
 import * as convert from "../utils/convert";
-import * as crypto from "../utils/crypto";
-import {ISigner, ITxJSON} from "../../interfaces";
-import Binary from "../Binary";
+import {keyTypeId} from "../utils/crypto";
+import {ITxJSON} from "../../interfaces";
 
 const DEFAULT_FEE = 100000000;
 const DEFAULT_VERSION = 3;
@@ -19,11 +18,11 @@ export default class Burn extends Transaction {
 		this.amount = amount;
 	}
 	private toBinaryV3(): Uint8Array {
-		return concatUint8Arrays(
+		return concatByteArray(
 			Uint8Array.from([this.type, this.version]),
-			Uint8Array.from(crypto.strToBytes(this.chainId)),
+			Uint8Array.from(strToBytes(this.chainId)),
 			Uint8Array.from(convert.longToByteArray(this.timestamp)),
-			Uint8Array.from([crypto.keyTypeId(this.senderKeyType)]),
+			Uint8Array.from([keyTypeId(this.senderKeyType)]),
 			base58.decode(this.senderPublicKey),
 			Uint8Array.from(convert.longToByteArray(this.fee)),
 			Uint8Array.from(convert.longToByteArray(this.amount)),
