+++ src/transactions/DataEntry.ts
@@ -1,5 +1,5 @@
 import * as convert from "../utils/convert";
-import {concatUint8Arrays} from "../utils/concat";
+import {concatByteArray} from "../utils/byte-array";
 import Binary from "../Binary";
 import {IHash} from "../../interfaces";
 
@@ -21,7 +21,7 @@ export default class DataEntry {
 	public toBinary(): Uint8Array {
 		const keyBytes = Uint8Array.from(convert.stringToByteArray(this.key));
 
-		return concatUint8Arrays(
+		return concatByteArray(
 			Uint8Array.from(convert.shortToByteArray(keyBytes.length)),
 			keyBytes, this.valueToBinary()
 		);
@@ -30,19 +30,19 @@ export default class DataEntry {
 	private valueToBinary(): Uint8Array {
 		switch (this.type) {
 		case "integer":
-			return concatUint8Arrays(
+			return concatByteArray(
 				Uint8Array.from([0]),
 				Uint8Array.from(convert.integerToByteArray(this.value as number))
 			);
 		case "boolean":
-			return concatUint8Arrays(
+			return concatByteArray(
 				Uint8Array.from([1]),
 				Uint8Array.from([+(this.value as boolean)])
 			);
 		case "binary":
-			return concatUint8Arrays(Uint8Array.from([2]), this.value as Uint8Array);
+			return concatByteArray(Uint8Array.from([2]), this.value as Uint8Array);
 		case "string":
-			return concatUint8Arrays(
+			return concatByteArray(
 				Uint8Array.from([3]),
 				Uint8Array.from(convert.stringToByteArray(this.value as string))
 			);
