+++ src/transactions/Register.ts
@@ -1,8 +1,8 @@
 import Transaction from "./Transaction";
-import { concatUint8Arrays } from "../utils/concat";
+import {concatByteArray, strToBytes} from "../utils/byte-array";
 import base58 from "../libs/base58";
 import * as convert from "../utils/convert";
-import * as crypto from "../utils/crypto";
+import {keyTypeId} from "../utils/crypto";
 import { Account } from "../accounts";
 import {IPublicAccount, ITxJSON} from "../../interfaces";
 
@@ -31,9 +31,9 @@ export default class Register extends Transaction {
 
 	private accountsToBinary(): Uint8Array {
 		return this.accounts.reduce(
-			(binary: Uint8Array, account: IPublicAccount) => concatUint8Arrays(
+			(binary: Uint8Array, account: IPublicAccount) => concatByteArray(
 				binary,
-				Uint8Array.from([crypto.keyTypeId(account.keyType)]),
+				Uint8Array.from([keyTypeId(account.keyType)]),
 				base58.decode(account.publicKey)
 			),
 			new Uint8Array()
@@ -41,11 +41,11 @@ export default class Register extends Transaction {
 	}
 
 	private toBinaryV3(): Uint8Array {
-		return concatUint8Arrays(
+		return concatByteArray(
 			Uint8Array.from([this.type, this.version]),
-			Uint8Array.from(crypto.strToBytes(this.chainId)),
+			Uint8Array.from(strToBytes(this.chainId)),
 			Uint8Array.from(convert.longToByteArray(this.timestamp)),
-			Uint8Array.from([crypto.keyTypeId(this.senderKeyType)]),
+			Uint8Array.from([keyTypeId(this.senderKeyType)]),
 			base58.decode(this.senderPublicKey),
 			Uint8Array.from(convert.longToByteArray(this.fee)),
 			Uint8Array.from(convert.shortToByteArray(this.accounts.length)),
